+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added `layout` to button_group field
 - Added `message` to true_false field
 - Added `stylisedUi` to select field
+- Added `disabled` to supported fields
 
 ## 8.3.0